requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pydantic==2.10.3
jinja2==3.1.4
pandas==2.2.3
openpyxl==3.1.5

authlib==1.3.2
httpx==0.28.1
itsdangerous==2.2.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.12
passlib[bcrypt]==1.7.4
